Refactored and moved logic of `package_handler.js`
